Refactored agent init to agent.cpp and renamed init_soar to run_soar
