+++ Core/SoarKernel/SoarKernel.cxx
@@ -3,9 +3,9 @@
 #include "src/decision_process/consistency.cpp"
 #include "src/decision_process/decide.cpp"
 #include "src/decision_process/decision_manipulation.cpp"
-#include "src/decision_process/init_soar.cpp"
 #include "src/decision_process/osupport.cpp"
 #include "src/decision_process/rete.cpp"
+#include "src/decision_process/run_soar.cpp"
 #include "src/episodic_memory/episodic_memory.cpp"
 #include "src/explain/explain_actions.cpp"
 #include "src/explain/explain_chunks.cpp"
